biodb/requirements.txt
Django>=1.8.1,<=1.8.14
django-userena>=2.0.0
django_tables2>=1.2.9
django-ckeditor>=5.1.1
django-simple-history>=1.8.1
django-bootstrap3>=7.0
django-watson>=1.3.0
django-guardian>=1.4.
easy-thumbnails>=2.3
pillow>=3.4